worker/requirements.txt
redis
pika
python-dotenv
pytest
timeout-decorator
